pyproject.toml
[tool.poetry]
name = "tubeatlas"
version = "0.1.0"
description = "Advanced YouTube Knowledge Graph & RAG Platform for transcript analysis and insight generation"
authors = ["Nicolas Schneider <nicolas.schneider@fh-bielefeld.de>"]
readme = "README.md"
packages = [{include = "tubeatlas", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"
youtube-transcript-api = "^0.6.0"
google-api-python-client = "^2.100.0"
langchain = "^0.1.0"
openai = "^1.0.0"
celery = {extras = ["redis"], version = "^5.3.0"}
redis = "^5.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.0"
tiktoken = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21.0"
coverage = "^7.0"
black = "^23.0"
isort = "^5.0"
flake8 = "^6.0"
mypy = "^1.0"
pre-commit = "^3.0"

# Tool configurations
detect-secrets = "^1.5.0"
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
src_paths = ["src", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
